// projects/aql-builder/src/app/shared/shared.enums.ts
/* Copyright 2021 Better Ltd (www.better.care)
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export enum AppLogo {
  AQL = 'logo-aql',
}

export enum NavigationIcons {
  HISTORY = 'nav-history',
  VIEWS = 'nav-views',
  PREDEFINED = 'nav-predefined',
  TEMPLATES = 'nav-templates',
  PUZZLE = 'nav-puzzle',
}

export enum InputType {
  DECIMAL = 'DECIMAL',
  CODED_TEXT = 'CODED_TEXT',
  TEXT = 'TEXT',
  INTEGER = 'INTEGER',
  BOOLEAN = 'BOOLEAN',
  DATE = 'DATE',
  TIME = 'TIME',
  DATETIME = 'DATETIME'
}